Added Linux versions of AssociatedType and GenericTypeParameter
