+++ SourceryRuntime/Sources/AST/TypeName/Generic.swift
@@ -80,74 +80,3 @@ public final class GenericType: NSObject, SourceryModelWithoutDescription, Diffa
 
 // sourcery:end
 }
-
-/// Descibes Swift generic type parameter
-#if canImport(ObjectiveC)
-@objcMembers
-#endif
-public final class GenericTypeParameter: NSObject, SourceryModel, Diffable {
-
-    /// Generic parameter type name
-    public var typeName: TypeName
-
-    // sourcery: skipEquality, skipDescription
-    /// Generic parameter type, if known
-    public var type: Type?
-
-    /// :nodoc:
-    public init(typeName: TypeName, type: Type? = nil) {
-        self.typeName = typeName
-        self.type = type
-    }
-
-    /// :nodoc:
-    override public var description: String {
-        var string = "\(Swift.type(of: self)): "
-        string += "typeName = \(String(describing: self.typeName))"
-        return string
-    }
-
-    public func diffAgainst(_ object: Any?) -> DiffableResult {
-        let results = DiffableResult()
-        guard let castObject = object as? GenericTypeParameter else {
-            results.append("Incorrect type <expected: GenericTypeParameter, received: \(Swift.type(of: object))>")
-            return results
-        }
-        results.append(contentsOf: DiffableResult(identifier: "typeName").trackDifference(actual: self.typeName, expected: castObject.typeName))
-        return results
-    }
-
-    public override var hash: Int {
-        var hasher = Hasher()
-        hasher.combine(self.typeName)
-        return hasher.finalize()
-    }
-
-    /// :nodoc:
-    public override func isEqual(_ object: Any?) -> Bool {
-        guard let rhs = object as? GenericTypeParameter else { return false }
-        if self.typeName != rhs.typeName { return false }
-        return true
-    }
-
-// sourcery:inline:GenericTypeParameter.AutoCoding
-
-        /// :nodoc:
-        required public init?(coder aDecoder: NSCoder) {
-            guard let typeName: TypeName = aDecoder.decode(forKey: "typeName") else { 
-                withVaList(["typeName"]) { arguments in
-                    NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: arguments)
-                }
-                fatalError()
-             }; self.typeName = typeName
-            self.type = aDecoder.decode(forKey: "type")
-        }
-
-        /// :nodoc:
-        public func encode(with aCoder: NSCoder) {
-            aCoder.encode(self.typeName, forKey: "typeName")
-            aCoder.encode(self.type, forKey: "type")
-        }
-
-// sourcery:end
-}
