+++ SourceryRuntime/Sources/AST/TypeName/GenericTypeParameter.swift
@@ -1,22 +1,19 @@
 #if canImport(ObjectiveC)
 import Foundation
 
-/// Describes Swift AssociatedType
+/// Descibes Swift generic type parameter
 @objcMembers
-public final class AssociatedType: NSObject, SourceryModel {
-    /// Associated type name
-    public let name: String
+public final class GenericTypeParameter: NSObject, SourceryModel, Diffable {
 
-    /// Associated type type constraint name, if specified
-    public let typeName: TypeName?
+    /// Generic parameter type name
+    public var typeName: TypeName
 
     // sourcery: skipEquality, skipDescription
-    /// Associated type constrained type, if known, i.e. if the type is declared in the scanned sources.
+    /// Generic parameter type, if known
     public var type: Type?
 
     /// :nodoc:
-    public init(name: String, typeName: TypeName? = nil, type: Type? = nil) {
-        self.name = name
+    public init(typeName: TypeName, type: Type? = nil) {
         self.typeName = typeName
         self.type = type
     }
@@ -24,57 +21,52 @@ public final class AssociatedType: NSObject, SourceryModel {
     /// :nodoc:
     override public var description: String {
         var string = "\(Swift.type(of: self)): "
-        string += "name = \(String(describing: self.name)), "
         string += "typeName = \(String(describing: self.typeName))"
         return string
     }
 
     public func diffAgainst(_ object: Any?) -> DiffableResult {
         let results = DiffableResult()
-        guard let castObject = object as? AssociatedType else {
-            results.append("Incorrect type <expected: AssociatedType, received: \(Swift.type(of: object))>")
+        guard let castObject = object as? GenericTypeParameter else {
+            results.append("Incorrect type <expected: GenericTypeParameter, received: \(Swift.type(of: object))>")
             return results
         }
-        results.append(contentsOf: DiffableResult(identifier: "name").trackDifference(actual: self.name, expected: castObject.name))
         results.append(contentsOf: DiffableResult(identifier: "typeName").trackDifference(actual: self.typeName, expected: castObject.typeName))
         return results
     }
 
     public override var hash: Int {
         var hasher = Hasher()
-        hasher.combine(self.name)
         hasher.combine(self.typeName)
         return hasher.finalize()
     }
 
     /// :nodoc:
     public override func isEqual(_ object: Any?) -> Bool {
-        guard let rhs = object as? AssociatedType else { return false }
-        if self.name != rhs.name { return false }
+        guard let rhs = object as? GenericTypeParameter else { return false }
         if self.typeName != rhs.typeName { return false }
         return true
     }
 
-// sourcery:inline:AssociatedType.AutoCoding
+// sourcery:inline:GenericTypeParameter.AutoCoding
 
         /// :nodoc:
         required public init?(coder aDecoder: NSCoder) {
-            guard let name: String = aDecoder.decode(forKey: "name") else { 
-                withVaList(["name"]) { arguments in
+            guard let typeName: TypeName = aDecoder.decode(forKey: "typeName") else {
+                withVaList(["typeName"]) { arguments in
                     NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: arguments)
                 }
                 fatalError()
-             }; self.name = name
-            self.typeName = aDecoder.decode(forKey: "typeName")
+             }; self.typeName = typeName
             self.type = aDecoder.decode(forKey: "type")
         }
 
         /// :nodoc:
         public func encode(with aCoder: NSCoder) {
-            aCoder.encode(self.name, forKey: "name")
             aCoder.encode(self.typeName, forKey: "typeName")
             aCoder.encode(self.type, forKey: "type")
         }
+
 // sourcery:end
 }
 #endif
